build: add AIX support
We have to upgrade the net-snmp in PE6, because of the OpenSSL 1.1.1
upgrade.

On AIX we have to build with ksh, or we will get a lot of exotic errors,
like not escaping libtool commands correctly, or not finding the print
command.

./configure SHELL=ksh does this job, except we have SHELL=/bin/bash
burnt in in our environment via MAKEFLAGS, when building with ZBS.
That needs to be removed in case of AIX.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>